python/valve-source/requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
docopt==0.6.2
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
gunicorn==19.8.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
monotonic==1.5
python-valve==0.2.1
requests==2.18.4
six==1.11.0
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
